Remove code paths only used in non-experimental mode
