+++ .gitignore
@@ -19,7 +19,6 @@ venv
 
 
 # Frontend are downloaded on release
-custom_components/hacs/hacs_frontend
 custom_components/hacs/hacs_frontend_experimental
 
 # Translation files
