+++ custom_components/hacs/frontend.py
@@ -30,9 +30,6 @@ if TYPE_CHECKING:
 async def async_register_frontend(hass: HomeAssistant, hacs: HacsBase) -> None:
     """Register the frontend."""
 
-    # Setup themes endpoint if needed
-    await hacs.async_setup_frontend_endpoint_themes()
-
     # Register frontend
     if hacs.configuration.dev and (frontend_path := os.getenv("HACS_FRONTEND_DIR")):
         hacs.log.warning(
