+++ custom_components/hacs/repositories/theme.py
@@ -38,8 +38,6 @@ class HacsThemeRepository(HacsRepository):
         except BaseException:  # lgtm [py/catch-base-exception] pylint: disable=broad-except
             pass
 
-        await self.hacs.async_setup_frontend_endpoint_themes()
-
     async def validate_repository(self):
         """Validate."""
         # Run common validation steps.
