+++ custom_components/hacs/utils/configuration_schema.py
@@ -10,4 +10,3 @@ NETDAEMON = "netdaemon"
 COUNTRY = "country"
 DEBUG = "debug"
 RELEASE_LIMIT = "release_limit"
-EXPERIMENTAL = "experimental"
