+++ custom_components/hacs/websocket/__init__.py
@@ -111,7 +111,7 @@ async def hacs_info(
                 "debug": hacs.configuration.debug,
                 "dev": hacs.configuration.dev,
                 "disabled_reason": hacs.system.disabled_reason,
-                "experimental": hacs.configuration.experimental,
+                "experimental": True,
                 "has_pending_tasks": hacs.queue.has_pending_tasks,
                 "lovelace_mode": hacs.core.lovelace_mode,
                 "stage": hacs.stage,
